+++ src/model/vectors/interfaces/Force.java
@@ -3,5 +3,5 @@ package model.vectors.interfaces;
 import model.vectors.Vector;
 
 public interface Force {
-    void apply(Vector vector, int deltaTime);
+    void apply(Vector vector, double deltaTime);
 }
